ci: stop testing against boulder
Pebble aims to be the standard ACME client test server.
This is good enough for this project, too. The protocol differences
between both are minimal or at least not relevant here.

However, maintaining boulde in CI is difficult - after all it is more
complex as it goal is to run in production.